Move everything into packages
