+++ app/src/androidTest/java/com/github/houseorganizer/houseorganizer/SettingsActivityTest.java
@@ -6,6 +6,8 @@ import static org.hamcrest.Matchers.is;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
+import com.github.houseorganizer.houseorganizer.activity.SettingsActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
