+++ app/src/main/java/com/github/houseorganizer/houseorganizer/house/HouseModel.java
@@ -1,4 +1,4 @@
-package com.github.houseorganizer.houseorganizer;
+package com.github.houseorganizer.houseorganizer.house;
 
 public class HouseModel {
     private String name;
