+++ app/src/main/java/com/github/houseorganizer/houseorganizer/shop/ShopItem.java
@@ -1,4 +1,4 @@
-package com.github.houseorganizer.houseorganizer;
+package com.github.houseorganizer.houseorganizer.shop;
 
 import androidx.annotation.NonNull;
