+++ app/src/main/java/com/github/houseorganizer/houseorganizer/user/User.java
@@ -1,4 +1,4 @@
-package com.github.houseorganizer.houseorganizer;
+package com.github.houseorganizer.houseorganizer.user;
 
 public abstract class User {
     abstract public String name();
