+++ app/src/main/java/com/github/houseorganizer/houseorganizer/util/BiViewHolder.java
@@ -1,4 +1,4 @@
-package com.github.houseorganizer.houseorganizer;
+package com.github.houseorganizer.houseorganizer.util;
 
 import android.view.View;
