+++ app/src/main/res/layout/activity_create_household.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CreateHouseholdActivity">
+    tools:context=".house.CreateHouseholdActivity">
 
     <EditText
         android:id="@+id/editTextHouseholdName"
