+++ app/src/main/res/layout/activity_house_selection.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HouseSelectionActivity">
+    tools:context=".house.HouseSelectionActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/housesView"
