+++ app/src/test/java/com/github/houseorganizer/houseorganizer/ShopItemUnitTest.java
@@ -4,6 +4,8 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 
+import com.github.houseorganizer.houseorganizer.shop.ShopItem;
+
 import org.junit.Test;
 
 public class ShopItemUnitTest {
